Implementación del sistema de mensajería: Se crea entidad Mensaje que relaciona usuarios con alertas/mensajes y que permite que los usuarios eliminen los mensajes que no les interesan. La creación de Alertas genera envío de correo.

+++ src/main/java/es/mira/progesin/persistence/repositories/IAlertaRepository.java
@@ -3,10 +3,13 @@ package es.mira.progesin.persistence.repositories;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
-
 import es.mira.progesin.persistence.entities.Alerta;
+import es.mira.progesin.persistence.entities.User;
 
-public interface IAlertaRepository extends CrudRepository<Alerta, Integer> {
+public interface IAlertaRepository extends CrudRepository<Alerta, Long> {
 
 	List<Alerta> findByFechaBajaIsNull();
+	
+
+
 }
